requirements.txt
fastapi
pyjwt
cryptography
uvicorn
pytest
